% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{givenness}
\alias{givenness}
\title{Givenness illusion data illustrating the "Williams Effect"}
\format{A data frame with 13 variables:
\describe{
\item{\code{condition}}{experimental condition}
\item{\code{item}}{item set number}
\item{\code{text}}{transcription of soundfiles}
\item{\code{participant}}{participant identifier}
\item{\code{conditionLabel}}{label of experimental condition}
\item{\code{npType}}{pronoun or full NP?}
\item{\code{voice}}{active or passive in second sentence?}
\item{\code{order}}{trial number in experiment}
\item{\code{acoustics}}{acoustic measures of final word}
\item{\code{stressshift}}{perceptual annotation whether final word was deaccented (and hence prominence shifted to earlier word)}
}}
\usage{
givenness
}
\description{
When we speak, some words have more emphasis than others: this is an aspect of prosody that linguists call *prosodic prominence*. Words that encode new or contrastive are likely to be prominent (e.g., they are likely to carry a pitch accent), words that encode old or repeated information are likely to be reduced and hence not prominent (e.g, they are unlikely to carry an accent)---unless they are contrastive. This experiment is an illustration of the 'Williams Effect': In English, a phrase-final word cannot be accented if a previous phrase ended in a homophonous accented word, even if it encodes contrastive information (e.g. the following sentence sounds odd if the final 'Betsy' is accented: 'Daniel followed Betsy, then Daniel was followed by Betsy.'). This suggests that speakers treat them as if they were merely contextually given and not contrastive, a 'givenness illusion'. For more information about contrast in general and the experiment in particular, please take a look at the references.
}
\references{
Wagner, M. (2012). A givenness illusion. Language and Cognitive Processes, 27(10):1433– 1458 (\href{http://dx.doi.org/10.1080/01690965.2011.607713}{doi})

Wagner, M. (2018). A givenness illusion. OSF project. \href{https://osf.io/r4j2w/}{project}
}
\keyword{datasets}
