% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tapping}
\alias{tapping}
\title{tapping}
\format{A data frame with 13 variables:
\describe{
\item{\code{participant}}{participant number}
\item{\code{speechrate}}{fast vs. slow}
\item{\code{item}}{number of item set}
\item{\code{syntax}}{intransitive vs. transitive}
\item{\code{tapped}}{annotation of whether word-final stop was tapped}
\item{\code{released}}{annotation of whether word-final stop was released}
\item{\code{vowelduration}}{duration of vowel preceding word-final stop}
\item{\code{segmentAfterDuration}}{duration of word-final stop}
\item{\code{pause}}{duration of pause after word-final stop}
\item{\code{trial}}{trial number in production experiment}
\item{\code{vowel}}{identity of vowel before word-final stop}
\item{\code{file}}{name of soundfile}
\item{\code{text}}{transcription of soundfile}
}}
\usage{
tapping
}
\description{
In North American English, the sounds [t] and [d] can sometimes be optionally pronounced as [ɾ] (a "tap") if followed by a vowel. This data set is from a production experiment in which two factors were manipulated: \code{speech rate} (speakers were instructed to take fast or to talk slowly) and \code{syntax} (instransitive: 'If you plit, Alice will be mad' vs. transitive: 'If you plit Alice, Fred will be mad. The data were annotated for whether the speaker tapped the consonant of interest (plit vs. pliɾ). Also, acoustic measures were extracted. See the references for further detail.
}
\references{
Kilbourn-Ceron, O. (2017). Speech production planning affects variation in external sandhi. PhD thesis, McGill University. \href{https://ling.auf.net/lingbuzz/003601}{thesis}

Kilbourn-Ceron, O., Wagner, M., and Clayards, M. (2017). The effect of production planning locality on external sandhi: A study in /t/. In Proceedings of the 52nd Annual Meeting of the Chicago Linguistic Society, pages 313–326. \href{https://ling.auf.net/lingbuzz/003119}{paper}

Kilbourn-Ceron, O., Wagner, M., and Clayards, M. (2017). The effect of production planning locality on external sandhi: A study in /t/. OSF Project \href{https://osf.io/8rjxu/}{project}
}
\keyword{datasets}
