% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trueAlternatives}
\alias{trueAlternatives}
\title{trueAlternatives}
\format{A data frame with 9 variables:
\describe{
 \item{\code{participant}}{id of speaker}
 \item{\code{item}}{number of item set}
 \item{\code{context}}{context speaker listened to before producting target sentence}
 \item{\code{rpitch}}{difference in semitones between maximum pitch on adjective and noun}
 \item{\code{rintensity}}{difference in dB between maximum intensity on adjective and noun}
 \item{\code{prominence}}{Annotation of whether prominence was shifted to adjective}
 \item{\code{trial}}{trial # duration experiment}
 \item{\code{text}}{Target text speaker was asked to respond with}
 }}
\usage{
trueAlternatives
}
\description{
Placing contrastive prosodic prominence on a word requires a contrsative alternative. But what counts as an 'alternative'? This dataset from a production experiment illustrates that at least when it comes to contrast marking on modifying adjectives, alternatives have to be mutually exclusive. See there references for further detail
}
\references{
Wagner, M. (2016). Information structure and production planning. In F ́ery, C. and Ishihara, S., editors, The Oxford Handbook of Information Structure, pages 541–561. Oxford University Press. \href{https://ling.auf.net/lingbuzz/002510}{preprint} \href{https://ling.auf.net/lingbuzz/002510}{http://dx.doi.org/10.1093/oxfordhb/9780199642670.013.39}

Wagner, M. (2018). Information structure and production planning. OSF Project \href{https://osf.io/8rjxu/}{project}
}
\keyword{datasets}
